+++ src/sf2_theme/__init__.py
@@ -2,6 +2,6 @@
 
 from typing import Final
 
-__version__: Final = "1.0.0"
+__version__: Final = "1.0.1"
 
 __all__ = ["__version__"]